Проверка 3-х веток (действий). Исправлены ошибки. 1 ветка (добавление нового элемента) – элементы не добавляются! При повторной загрузке элементы появились – странно. Оказывается, в бд все прекрасно добавляется в EditActivity, но id студента не передается из-за несоответствия типов. Int != long. Эта ошибка идет из объявления бд – id элемента, значение, возвращаемое из методов update и insert имеет тип long. Исправлено. 2 ветка (редакция по щелчку) – информацию о студенте загрузил правильно, но выдал ошибку при переходе в MainActivity. Оказывается, неправильно выполняется операция обновления бд, третьим аргументом идет: “id = ?”, а не просто “id”. Исправлено. 2 ветка (редакция по контекстному меню) – работает правильно. 3 ветка (удаление элемента) – глюк – элементу, под удаляемым, присваивается чужая фотография? Однако, при повторной загрузке такого не происходит. Интересно, что такого не происходит если у нижнего элемента есть картинка. Исправлено вводом картинки по умолчанию.

+++ app/src/main/java/com/example/artemij/ppplus/NStudent.java
@@ -32,7 +32,7 @@ public class NStudent implements Parcelable {
         out.writeString(chap.getName());
         out.writeString(chap.getNickName());
         out.writeInt(chap.getPlusAmount());
-        out.writeInt(chap.getID());
+        out.writeLong(chap.getID());
         out.writeString(chap.getStringImageUri());
     }
 
@@ -44,7 +44,7 @@ public class NStudent implements Parcelable {
         chap.setName(in.readString());
         chap.setNickName(in.readString());
         chap.setPlusAmount(in.readInt());
-        chap.setID(in.readInt());
+        chap.setID(in.readLong());
         chap.setUri(in.readString());
 
     }
